requirements.txt
fastapi
pydantic
yt-dlp
python-dotenv
youtube-transcript-api
langchain>=0.1.0
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
transformers
sentence-transformers
huggingface-hub
pymongo
passlib[bcrypt]
python-jose
uvicorn
streamlit
requests